New "-X exclusions-file" option for mconfig
Packages can use this option to list symbols that shouldn't bring in the
corresponding units. For example, Perl need not provide support for BSD
index(3) as an alternative to C89 strchr(3), but "index" is the name of a
Perl builtin, so that string in the source files is misunderstood by
metaconfig as an attempt to use the BSD function.

With this change, Perl can deal with this situation by adding "index" (and
"rindex") to an exclusion list.